fix the connectivity of backend and frontend

+++ backend/src/main/java/com/example/IT342_G1_Maturan/dto/AuthResponse.java
@@ -2,17 +2,17 @@ package com.example.IT342_G1_Maturan.dto;
 
 public class AuthResponse {
     private String token;
-    private Integer userId;
+    private Long userId;
 
     public AuthResponse() {}
 
-    public AuthResponse(String token, Integer userId) {
+    public AuthResponse(String token, Long userId) {
         this.token = token;
         this.userId = userId;
     }
 
     public String getToken() { return token; }
     public void setToken(String token) { this.token = token; }
-    public Integer getUserId() { return userId; }
-    public void setUserId(Integer userId) { this.userId = userId; }
+    public Long getUserId() { return userId; }
+    public void setUserId(Long userId) { this.userId = userId; }
 }
